перенес методы актеров,  в moviesDao
